// src/enum/view_permission.enum.ts
export enum ViewPermission {
  DASHBOARD_VIEW = 'dashboard.view',
  PRODUCTS_VIEW = 'products.view',
  CATEGORIES_VIEW = 'categories.view',
  STOCK_MOVEMENTS_VIEW = 'stock_movements.view',
  PURCHASE_ORDERS_VIEW = 'purchase_orders.view',
  SUPPLIERS_VIEW = 'suppliers.view',
  USERS_VIEW = 'users.view',
  BUSINESS_VIEW = 'business.view',
  PROFILE_VIEW = 'profile.view',
  REPORTS_VIEW = 'REPORTS_VIEW',
  STOCKS_VIEW = 'STOCKS_VIEW',
  STORES_VIEW = 'STORES_VIEW',
  AUDIT_LOGS_VIEW = 'AUDIT_LOGS_VIEW',
  PRODUCT_SOURCES_VIEW = 'PRODUCT_SOURCES_VIEW',
}